updating should work properly now

+++ src/bin/update-arnix.sh
@@ -66,8 +66,10 @@ cp -a arnix-bootstrap.sha1sum.txt /arnix/arnix-bootstrap.sha1sum.txt
 
 cd /arnix/etc
 for i in *; do
-    [[ $i = '*.arnixnew' ]] && rm -rf $i && continue
-    [[ $i = '*.sha1sum.txt' ]] && continue
+    [[ "$i" = *'.arnixnew' ]] && rm -rf $i && continue
+    [[ "$i" = *'.sha1sum.txt' ]] && continue
+
+    [ ! -e /arnix/merge/etc/$i ] && rm -f /arnix/etc/$i && continue
     
     # Check if the original checksum still matches
     [ -e $i.sha1sum.txt ] && \
@@ -85,6 +87,8 @@ done
 
 cd /arnix/bin
 for i in *; do
+    [[ "$i" = *'.sha1sum.txt' ]] && continue
+    
     [ ! -e /arnix/merge/bin/$i ] && rm -f /arnix/bin/$i
     
     cp -rf /arnix/merge/bin/$i $i
